# contracts/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plant Doctor RAG API
  version: 0.2.0
  description: |
    식물 주치의 AI MVP용 프론트-백엔드 공유 API 계약 문서입니다.
    
    [업데이트 히스토리]
    * v0.2.0: Supabase pgvector 클라우드 DB 연동 및 RAG 상담 이력 메시지 스키마 정합 반영 완료.
    * v0.1.0: RAG 상담 및 이력 조회 API 엔드포인트 초안 수립.
servers:
  - url: http://localhost:8000
    description: Local backend
security:
  - bearerAuth: []
paths:
...
  /api/v1/gardens:
    get:
      summary: List current user's gardens
      responses:
        "200":
          description: Garden list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Garden"
    post:
      summary: Create a garden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GardenCreate"
      responses:
        "201":
          description: Created garden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Garden"
  /api/v1/gardens/{gardenId}:
    patch:
      summary: Update a garden
      parameters:
        - $ref: "#/components/parameters/GardenId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GardenUpdate"
      responses:
        "200":
          description: Updated garden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Garden"
        "404":
          description: Garden not found or access denied
    delete:
      summary: Delete a garden and unassign its plants
      parameters:
        - $ref: "#/components/parameters/GardenId"
      responses:
        "204":
          description: Garden deleted successfully
        "404":
          description: Garden not found or access denied
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    GardenId:
      name: gardenId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    GardenCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 40
        location:
          type: string
          maxLength: 60
        description:
          type: string
          maxLength: 80
        sunlight:
          type: string
          maxLength: 40
        soilType:
          type: string
          maxLength: 40
    GardenUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 40
        location:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        sunlight:
          type: string
          nullable: true
        soilType:
          type: string
          nullable: true
    Garden:
      allOf:
        - $ref: "#/components/schemas/GardenCreate"
        - type: object
          required: [id, plantCount, createdAt]
          properties:
            id:
              type: string
              format: uuid
            plantCount:
              type: integer
              minimum: 0
            imageUrl:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
...